requirements.txt
Django==1.7
argparse==1.2.1
django-push-notifications==1.2.0
django-tastypie==0.12.1
jsonfield==1.0.0
parse-rest==0.2.20141004
python-dateutil==2.2
python-mimeparse==0.1.4
requests==2.4.1
six==1.8.0
twill==0.9.1
wsgiref==0.1.2
